+++ src/main/java/bestie/command/InvalidCommand.java
@@ -5,7 +5,13 @@ import bestie.TaskList;
 import bestie.Ui;
 
 public class InvalidCommand extends Command {
-
+    /**
+     * Prints message on the console to show that user has entered an invalid command.
+     *
+     * @param tasks User's list of tasks.
+     * @param ui User Interface object that executes printing of message to console.
+     * @param storage Loads list of tasks from file and writes tasks to the bestie.txt file.
+     */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         ui.invalidCommand();
