+++ src/main/java/bestie/command/ListCommand.java
@@ -5,7 +5,13 @@ import bestie.TaskList;
 import bestie.Ui;
 
 public class ListCommand extends Command {
-    // display all the tasks in the list
+    /**
+     * Displays all the tasks in the user's task list.
+     *
+     * @param tasks User's list of tasks.
+     * @param ui User Interface object that executes printing of message to console.
+     * @param storage Loads list of tasks from file and writes tasks to the bestie.txt file.
+     */
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         ui.displayTasks(tasks.getTasks());
