+++ src/main/java/bestie/task/Deadline.java
@@ -1,5 +1,4 @@
 package bestie.task;// import libraries necessary for bestie to understand the dates
-import bestie.Task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
