+++ src/main/java/bestie/task/Event.java
@@ -1,7 +1,5 @@
 package bestie.task;
 
-import bestie.Task;
-
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
